Create Game model
Game model is a class that has a board (of type array) as a class variable.
A new Game initializes a new 6x7 array as a board.

Game class currently has 3 methods:
self.get_board - returns the current state of the board
sef.place_move - given the player's value and the player's move
represented by (x,y) coordinates, the board's (x,y) values are update with
the player's value
self.is_winner - given the player's current move, it checks if a
winning combination is made.